PORTALS-2771 - primary key equality check to determine row selection
PORTALS-2771 references a case where there is no `rowId`, which breaks 'selected' checkbox state. Add a new helper to determine this state. When defined, we should reference the `rowSelectionPrimaryKey` (so add it for the corresponding EL Portal Cohort Builder case). If there's no specified primary key, gracefully fall back to using rowId or the `id` column when it is appropriate, using deep comparison of Row objects as a last resort.

One goal that is considered, though not yet attained, is to allow users to maintain a cohort selection between sessions. Since data in the portal may change over time, using a primary key is more resilient than using deep equality. However, this is still not perfect; we don't save column models, so this would still break if the index of the primary key column(s) were to change.