SWC-6552 - Fix UserOrTeamBadge showing stale data
`UserOrTeamBadge` provides two 'interfaces' so the userGroupHeader data can be either provided or fetched. This bug only occurred when data was fetched, but the faulty logic existed to manage both scenarios.

The data was stored in `useState`. Due to incorrect logic in an effect, the data was NOT updated when the fetched data changed due to the `principalId` changing.

We can remove the effect and just pick the correct object inline and the bug is resolved.