Add error lint rule "react/no-object-type-as-default-prop"
This ESLint rule prevents us from specifying new objects as default props in components, which are re-created on each new render cycle, which can lead to performance issues. Note that default primitive values (which are effectively compared by value by React) are unaffected.

I also added the `quiet` option to the root `lint` script, which will cause the command to only output errors instead of errors and warnings, which is more useful in CI.