PORTALS-3420
Refactor to create simple hook wrappers to get row selection Jotai atom values and the ReactQuery result for queryMetadata. This refactor improve testability by simplifying the process to create mocks. Without the wrappers, we would have to mock `useQuery` or `useAtomValue` and the mock implementation would have to add logic to handle the different atoms or query options.

With the wrapper, we can simply mock the wrapper without writing complex mock logic for external dependencies.