SWC-7467
- Support creating a Grid from a RecordSet ID
- Modify SynapseGrid imperative handle
   - Change initializeGrid to accept the CreateGridRequest
   - Add `loadExistingSession` to load from an existing session ID.

In SynapseWebClient, to support instantiating a GridSession from a RecordSet, we will move the Grid UI to its own page. This page will be parameterized by the grid's sessionId. To enable this, SynapseGrid needs to be able to accept the session ID as an argument to the imperative handle function.